requirements.txt
groq
python-dotenv
edge-tts
pygame
requests
chromadb
rich
ddgs
